+++ plugin/mpi/SLEPc.cpp
@@ -1,8 +1,9 @@
 //ff-c++-LIBRARY-dep: cxx11 slepc petsc [mkl|blas] hpddm mpi
 //ff-c++-cpp-dep:
 
-#define WITH_slepc
-#define  PETScandSLEPc 1
-#include "PETSc-code.hpp"
-#include "SLEPc-code.hpp"
-LOADFUNC(Init)
+static void Load_Init() {
+    if (mpirank == 0)
+        cerr << " ++ WARNING: SLEPc has been superseded by PETSc" << endl;
+}
+
+LOADFUNC(Load_Init)